List Branches of the moviesApp Repository

This function clones the 'moviesApp' GitHub repository, extracts the list of branches along with their last update time, and then returns this information as a tibble. The repository is cloned to a temporary directory, which is removed after the function executes.

Usage

list_apps(regex = NULL)

Arguments

regex

An optional character string containing a regular expression (regex) to filter branch names

Value

A tibble with three columns: 'source', 'branch_name', and 'updated', listing the source of the branch, the branch name, and the last updated date.

How it works

The function operates in several steps:

It saves the current working directory to revert back to it later

It clones the 'moviesApp' repository from GitHub into a temporary directory

It lists all branches of the cloned repository

It separates the branch names into source and branch_name

It cleans up by deleting the temporary directory and restoring the original working directory

Note that the function currently does not use the regex parameter to filter branches. This parameter can be implemented in future versions to allow branch filtering based on regular expressions.

Examples

list_apps()
#> # A tibble: 54 × 2
#>    branch              last_updated       
#>    <chr>               <dttm>             
#>  1 01_whole-app-game   2024-01-17 19:30:27
#>  2 02.1_shiny-app      2024-01-17 19:35:35
#>  3 02.2_movies-app     2024-01-17 19:37:19
#>  4 02.3_proj-app       2024-01-17 19:39:28
#>  5 03.1_description    2024-01-17 19:40:11
#>  6 03.2_rproj          2024-01-17 19:41:14
#>  7 03.3_create-package 2024-01-17 19:42:45
#>  8 04_devtools         2024-01-17 19:44:38
#>  9 05_roxygen2         2024-01-17 19:45:34
#> 10 06.1_pkg-exports    2024-01-17 19:46:25
#> # ℹ 44 more rows
list_apps(regex = "tests")
#> # A tibble: 6 × 2
#>   branch               last_updated       
#>   <chr>                <dttm>             
#> 1 11_tests-specs       2024-02-21 17:09:11
#> 2 12.1_tests-fixtures  2024-02-21 17:11:15
#> 3 12.2_tests-helpers   2024-02-21 17:12:29
#> 4 12.3_tests-snapshots 2024-02-21 17:31:28
#> 5 13_tests-modules     2024-02-21 17:33:30
#> 6 14_tests-system      2024-02-21 19:02:02